pyproject.toml
[project]
name = "voice-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "assemblyai>=0.37.0",
    "boto3>=1.37.19",
    "boto3-stubs>=1.37.29",
    "colorama>=0.4.6",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "groq>=0.20.0",
    "loguru>=0.7.3",
    "numpy>=2.0.2",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "pywav>=1.1",
    "requests>=2.32.3",
    "scipy>=1.13.1",
    "sounddevice>=0.5.1",
    "twilio>=9.5.1",
    "types-requests>=2.31.0.6",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]

[tool.mypy]
ignore_missing_imports = true
